third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/long-press-focuses-frame-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.activeElement == frame1 is true
PASS document.activeElement == frame2 is true
PASS document.activeElement == document.body is true
This test checks that long press to select text focuses the pressed frame.
 
